+++ demisto_sdk/commands/content_graph/objects/relationship.py
@@ -8,11 +8,13 @@ from demisto_sdk.commands.content_graph.objects.base_content import BaseContent
 
 class RelationshipData(BaseModel):
     relationship_type: RelationshipType
-    source: BaseContent
-    target: BaseContent
+
+    # These are the database ids of the relationships
+    source_id: int
+    target_id: int
 
     # this is the attribute we're interested in when querying
-    content_item: BaseContent
+    content_item_to: BaseContent
 
     is_direct: bool = True
 
@@ -31,11 +33,9 @@ class RelationshipData(BaseModel):
 
         return hash(
             (
-                self.source.database_id,
-                self.target.database_id,
+                self.source_id,
+                self.target_id,
                 self.relationship_type,
-                self.source.content_type,
-                self.target.content_type,
             )
         )